pyproject.toml
[project]
name = "agente-rh-com-memoria"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-huggingface>=1.2.1",
    "pypdf>=6.7.5",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.55.0",
]
